// packages/uikit-desk/src/view/DeskMonitor/types.ts
export interface DeviceInfo {
    walletAccountId?:string
    deviceId: string;
    name?: string;
    connected?: boolean;
    password?: string;
    passwordHash?: string;
    serverApi?: string;
    inputIsOpen?: boolean;
    mediaIsStart?: boolean;
    compressQuality?: number;
    delayPullEventMs?: number;
    delaySendImageDataMs?: number;
    screen?: {
        height: number;
        width: number;
        scale: number;
        dpi: number;
    };
}

export enum SWIPER_KIND {
    SWIPER_KIND_UP_DOWN = 10,
    SWIPER_KIND_UP_DOWN_SLOW = 11,
    SWIPER_KIND_LEFT_RIGHT = 20,
    SWIPER_KIND_LEFT_RIGHT_SLOW = 21
}

export enum WsCloseCode {}

export enum WsCloseCode {
    //@delete
    CLOSE_AND_DESTROY = 1001,
    //@delete
    CLOSE_AND_RECONNECT = 1002,

    WS_CLOSE_STOP_RECONNECT = 3001,
    WS_CLOSE_RECONNECT = 3002
}

export enum ErrCodes {
    DEVICE_NOT_EXISTS = 'DEVICE_NOT_EXISTS',
    PASSWORD_NOT_VALID = 'PASSWORD_NOT_VALID',
    SCREEN_IMAGE_NOT_EXISTS = 'SCREEN_IMAGE_NOT_EXISTS',
    SCREEN_IMAGE_NOT_UPDATE = 'SCREEN_IMAGE_NOT_UPDATE',
    EVENT_NOT_UPDATE = 'EVENT_NOT_UPDATE'
}
// ...
export interface CutAreaRect {
    start: { x: number; y: number };
    end: { x: number; y: number };
}

export interface RecAreaRect {
    start: { x: number; y: number };
    end: { x: number; y: number };
    maxVal:number
}

export interface XYWHProps {
    x: number;
    y: number;
    w: number;
    h: number;
}

export interface DeviceOptions {
    walletAccountId?:string;
    canCut?: boolean;
    tgUserId?: string;
    webviewTitle?: string;
    customHosts?: { host: string }[];
    setCustomHosts?: (v:{host:string}[])=>void;
    partitionId?: string;
    deleteTgSite?: (userId: string)=>void;
    canDelete?: boolean;
    headerColor?: PropTypes.Color | 'transparent' | 'error' | 'info' | 'success' | 'warning';
    isInfoPanel?: boolean,
    setIsInfoPanel?: (v: boolean)=>void,
    isSettingPanel?: boolean,
    setIsSettingPanel?: (v: boolean)=>void,
    setGlobalUpdatedAt?: (v: number)=>void;
    setUpdatedAt?: (v: number)=>void;
    setSnackbar: (v: string)=>void;
    isAdding?: boolean;
    optionsTab?: 'CUT_AREA' | string,
    setOptionsTab?: (v: string)=>void,
    webview?: any;
    isWebview?: boolean;
    logsUpdateAt: number;
    isLogged?: boolean;
    recognitionAreaRect: CutAreaRect[];
    setRecognitionAreaRect: (v: RecAreaRect[])=>void;
    cutAreaRect: CutAreaRect;
    setCutAreaRect: (v: CutAreaRect)=>void;
    ws?: WebSocket;
    isCutEnable: boolean;
    setIsCutEnable: (v: boolean)=>void;
    monitorScale: number;
    setMonitorScale: (v: number)=>void;
    errCode?: ErrCodes;
    auth?: (deviceId:string, password:string, serverApi:string)=>Promise<boolean>;
    deviceId: string;
    updatedAt?: number;
    setDeviceId?: (v: string)=>void;
    getDeviceInfo: (key?: keyof DeviceInfo, defaultVal?: any) => any;
    onChangeDeviceInfo?: (key: keyof DeviceInfo, value: any)=>void;
}